code style fixes

+++ binominalCoefficient.cpp
@@ -53,63 +53,63 @@ void buildPascal()
 // https://comeoncodeon.wordpress.com/2011/07/31/combination/
 
 /* This function calculates power of p in n! */
-int countFact(int n, int p)
+ll countFact(ll n, ll p)
 {
-    int k=0;
-    while (n>=p)
+    ll k = 0;
+    while (n >= p)
     {
-        k+=n/p;
-        n/=p;
+        k += n / p;
+        n /= p;
     }
     return k;
 }
- 
+
 /* This function calculates (a^b)%MOD */
-long long pow(int a, int b, int MOD)
+long long pow(ll a, ll b, ll MOD)
 {
-    long long x=1,y=a; 
-    while(b > 0)
+    long long x = 1, y = a;
+    while (b > 0)
     {
-        if(b%2 == 1)
+        if (b % 2 == 1)
         {
-            x=(x*y);
-            if(x>MOD) x%=MOD;
+            x = (x * y);
+            if (x > MOD) x %= MOD;
         }
-        y = (y*y);
-        if(y>MOD) y%=MOD; 
+        y = (y * y);
+        if (y > MOD) y %= MOD;
         b /= 2;
     }
     return x;
 }
- 
+
 /*  Modular Multiplicative Inverse
     Using Euler's Theorem
     a^(phi(m)) = 1 (mod m)
     a^(-1) = a^(m-2) (mod m) */
-long long InverseEuler(int n, int MOD)
+long long InverseEuler(ll n, ll MOD)
 {
-    return pow(n,MOD-2,MOD);
+    return pow(n, MOD - 2, MOD);
 }
- 
-long long factMOD(int n, int MOD)
+
+long long factMOD(ll n, ll MOD)
 {
-    long long res = 1; 
+    long long res = 1;
     while (n > 0)
     {
-        for (int i=2, m=n%MOD; i<=m; i++)
+        for (ll i = 2, m = n % MOD; i <= m; i++)
             res = (res * i) % MOD;
-        if ((n/=MOD)%2 > 0) 
+        if ((n /= MOD) % 2 > 0)
             res = MOD - res;
     }
     return res;
 }
- 
-long long C(int n, int r, int MOD)
+
+long long C(ll n, ll r, ll MOD)
 {
-    if (countFact(n, MOD) > countFact(r, MOD) + countFact(n-r, MOD))
+    if (countFact(n, MOD) > countFact(r, MOD) + countFact(n - r, MOD))
         return 0;
- 
+
     return (factMOD(n, MOD) *
-            ((InverseEuler(factMOD(r, MOD), MOD) * 
-            InverseEuler(factMOD(n-r, MOD), MOD)) % MOD)) % MOD;
+        ((InverseEuler(factMOD(r, MOD), MOD) *
+            InverseEuler(factMOD(n - r, MOD), MOD)) % MOD)) % MOD;
 }
